@@ -102,13 +17,56 @@
     "greenkeeperio-bot <support@greenkeeper.io>"
   ],
   "license": "MIT",
+  "homepage": "https://github.com/node-ebml/node-ebml#readme",
+  "repository": "github:node-ebml/node-ebml",
   "bugs": {
     "url": "https://github.com/node-ebml/node-ebml/issues"
   },
-  "homepage": "https://github.com/node-ebml/node-ebml#readme",
-  "lint-staged": {
-    "*.js": "eslint --fix --no-ignore",
-    "*.json": "prettier --write",
-    "README.md": "prettier --write"
+  "keywords": [
+    "ebml",
+    "webm",
+    "mkv",
+    "matroska",
+    "format"
+  ],
+  "main": "lib/ebml.js",
+  "module": "lib/ebml.esm.js",
+  "directories": {
+    "test": "__test__",
+    "lib": "lib"
+  },
+  "scripts": {
+    "test": "jest",
+    "lint": "eslint",
+    "build": "rollup --config",
+    "prepublishOnly": "NODE_ENV=production pnpm run build"
+  },
+  "dependencies": {
+  },
+  "devDependencies": {
+    "@antfu/eslint-config": "^4.10.1",
+    "@rollup/plugin-commonjs": "~28.0.3",
+    "@rollup/plugin-json": "~6.1.0",
+    "@rollup/plugin-node-resolve": "^16.0.1",
+    "@rollup/plugin-replace": "6.0.2",
+    "@rollup/plugin-terser": "0.4.4",
+    "@types/rollup-plugin-node-builtins": "^2.1.5",
+    "@types/rollup-plugin-node-globals": "^1.4.5",
+    "check-node-version": "^4.2.1",
+    "eslint": "^9.22.0",
+    "jest": "~29.7.0",
+    "jest-html-reporter": "^4.1.0",
+    "lodash.foreach": "^4.5.0",
+    "lodash.map": "^4.6.0",
+    "lodash.range": "^3.2.0",
+    "mocha": "~11.1.0",
+    "pnpm": "^10.6.5",
+    "rimraf": "^6.0.1",
+    "rollup": "^4.36.0",
+    "rollup-plugin-node-builtins": "^2.1.2",
+    "rollup-plugin-node-globals": "^1.4.0",
+    "typescript": "^5.8.2",
+    "unexpected": "^13.2.1",
+    "unexpected-date": "^2.0.0"
   }
 }
