@@ -1,7 +1,7 @@
-module.exports = {
+export default {
   collectCoverage: true,
   collectCoverageFrom: ['src/**/*.js'],
   coverageDirectory: './coverage',
   coverageReporters: ['json', 'lcov', 'text', 'html'],
   testURL: 'http://localhost/',
-};
+}
