@@ -1,16 +1,13 @@
-/* eslint-disable node/no-unsupported-features/es-syntax */
-import babel from '@rollup/plugin-babel';
-import builtins from 'rollup-plugin-node-builtins';
-import commonjs from '@rollup/plugin-commonjs';
-import globals from 'rollup-plugin-node-globals';
-import json from '@rollup/plugin-json';
-import resolve from '@rollup/plugin-node-resolve';
-import replace from '@rollup/plugin-replace';
-import { terser } from 'rollup-plugin-terser';
-import { defineConfig } from 'rollup';
+import process from 'node:process'
+import json from '@rollup/plugin-json'
+import resolve from '@rollup/plugin-node-resolve'
+import replace from '@rollup/plugin-replace'
+import terser from '@rollup/plugin-terser'
+import { defineConfig } from 'rollup'
+import builtins from 'rollup-plugin-node-builtins'
+import globals from 'rollup-plugin-node-globals'
 
 const plugins = [
-  babel({ exclude: 'node_modules/**', babelHelpers: 'bundled' }),
   resolve({
     browser: true,
     jsnext: true,
@@ -18,7 +15,6 @@ const plugins = [
     module: true,
     preferBuiltins: true,
   }),
-  commonjs(),
   builtins(),
   globals(),
   replace({
@@ -28,46 +24,18 @@ const plugins = [
     preventAssignment: true,
   }),
   json(),
-];
+]
 
-const sourcemap = process.env.SOURCE_MAPS || true;
-const globalOpts = {
-  stream: 'stream',
-};
+const sourcemap = process.env.SOURCE_MAPS || true
 
 export default defineConfig([
   {
     input: './src/ebml/index.js',
     output: [
       {
-        file: 'lib/ebml.js',
-        format: 'cjs',
-        sourcemap,
-      },
-      {
         file: 'lib/ebml.esm.js',
         format: 'esm',
         sourcemap,
-      },
-      {
-        file: 'lib/ebml.iife.js',
-        format: 'iife',
-        name: 'EBML',
-        sourcemap,
-        globals: globalOpts,
-      },
-      {
-        file: 'lib/ebml.amd.js',
-        format: 'amd',
-        name: 'EBML',
-        sourcemap,
-      },
-      {
-        file: 'lib/ebml.umd.js',
-        format: 'umd',
-        name: 'EBML',
-        sourcemap,
-        globals: globalOpts,
       },
     ],
     plugins,
@@ -76,32 +44,10 @@ export default defineConfig([
     input: './src/ebml/index.js',
     output: [
       {
-        file: 'lib/ebml.min.js',
-        format: 'cjs',
-      },
-      {
         file: 'lib/ebml.esm.min.js',
         format: 'esm',
-      },
-      {
-        file: 'lib/ebml.iife.min.js',
-        format: 'iife',
-        name: 'EBML',
-        globals: globalOpts,
-      },
-      {
-        file: 'lib/ebml.amd.min.js',
-        format: 'amd',
-        name: 'EBML',
-      },
-      {
-        file: 'lib/ebml.umd.min.js',
-        format: 'umd',
-        name: 'EBML',
-        globals: globalOpts,
       },
     ],
     plugins: [...plugins, terser()],
   },
-]);
-/* eslint-enable node/no-unsupported-features/es-syntax */
+])
