@@ -1,4 +1,4 @@
-export tools from './tools';
-export schema from './schema';
-export Decoder from './decoder';
-export Encoder from './encoder';
+export { default as Decoder } from './decoder'
+export { default as Encoder } from './encoder'
+export { default as Schema } from './schema'
+export { default as Tools } from './tools'
